// src/app/models/student.ts
export interface Student {
    student_code : number;
    student_name : string;
    class_student : string;
    major: string;
    year_of_study: number;
    company_id: number;
    mentor_id: number;
    lecturer_id: number;
    start_date: Date;
    end_date: Date;
    language: string;
    position: string;
    user_id : number;
    status: string;
}